+++ druid/src/widget/spinner.rs
@@ -72,6 +72,7 @@ impl<T: Data> Widget<T> for Spinner {
     fn lifecycle(&mut self, ctx: &mut LifeCycleCtx, event: &LifeCycle, _data: &T, _env: &Env) {
         if let LifeCycle::WidgetAdded = event {
             ctx.request_anim_frame();
+            ctx.request_paint();
         }
 
         if let LifeCycle::AnimFrame(interval) = event {
@@ -80,6 +81,7 @@ impl<T: Data> Widget<T> for Spinner {
                 self.t = 0.0;
             }
             ctx.request_anim_frame();
+            ctx.request_paint();
         }
     }